Perform nightly tally via DB query
Perform the nightly tally for an org by querying the DB instead
of holding all hosts and their facts in memory to perform the
calculations.

The query does the tally by targeting the host_tally_buckets
and summing based on the values for each bucket. It is using
the cores and sockets on the bucket itself because these fields
are normailzed before being stored, unlike the instance measurements.

The `LEGACY_NIGHTLY_TALLY_ENABLED` environment variable was added to
allow enabling the old legacy tally functionality, should we need to.

Legacy tally is being tested in its own class with the
intent that these tests will be removed once the new
tally logic has been tested and finalized for production.