Set virtual measurement type for non-RHEL when tallying
For virtual, non RHEL systems, we were only adding them
to the overall system totals, and no tracking them as
virtual. Because of this, the system list would display
the virtual systems, however they would not be reflected
in the counts in the graph.

This patch properly tracks them as virtual.

NOTE:
 This only applies to the LEGACY tally since the new
 approach of tallying via the DB will fix this issue
 automatically. Should we need to disable the new
 style tally, this fix will still be applied.