Suppress warning about shifting account number for empty string
To set the account number on the AccountUsageCalculation, we pull the
account number from the NormalizedFacts of the hosts.  The code is
written such that the account number is set when processing the first
host and that subsequent host processing will check to make sure the
account number is unchanged in the NormalizedFacts.  Any NormalizedFacts
object with a differing account number results in an
IllegalArgumentException.  A comment notes that "there could be stale
data in inventory with no account set".

There is an edge case where the first NormalizedFacts object has no
account set and the others do.  In that case, we would actually want to
update the account number with the more meaningful data.  This patch
alters InventoryAccountUsageController#collect() to only throw the
IllegalArgumentException if the AccountUsageCalculation account number
is non-empty, non-null, and non-whitespace, and we receive a differing
account number value from the NormalizedFacts object.  If the
AccountUsageCalculation account number is null, empty, or only
whitespace, it will be updated with the NormalizedFacts object account
number.

See https://issues.redhat.com/browse/SWATCH-868