Remove cast on instance_id in tally_instance_view
Our instance view was casting host.instance_id to a UUID, but,
host.instance_id on the hosts table isn't always a proper UUID (
i.e OpenShift Clusters).

When querying against the view, if the instance_id stored in the
host table is not a proper UUID, it would cause an error at the DB level.

This patch removes the cast on instance_id and updates the field to a
String in the view class.