Add contractEnabled flag to TagMetaData
TagMetaData now has a new 'contractEnabled' field that can
be set for a configured tag. By default it is set to false.

The contractEnabled field can only be set to true if the
billingModel field is 'PAYG'. If not, the application will
fail to start up, indicating a configuration error.

The BASILISK configuration has been updated to enable contracts.