Add Red Hat Partner Gateway API client to swatch-contracts
https://issues.redhat.com/browse/SWATCH-944

This adds the client, and its configuration to swatch-contracts.

I also defined the Partner Gateway URLs in quarkus profiles, and added
the `QUARKUS_PROFILE` template param/env var, so we won't need to
manually specify the URL for deployment.

It can still be overridden by setting the `ENTITLEMENT_GATEWAY_URL` env
var.

Supported quarkus profiles:
* `dev` (default when running locally w/ `quarkus dev` or
  `./gradlew :swatch-contracts:quarkusDev`)
* `qa`
* `stage`
* `prod` (default when running from the jar)

Testing
=======

The added integration test can be run via:

```
./gradlew swatch-contracts:test --tests=RhPartnerClientIntegrationTest
```

(it is also run with all the other tests).