Log unknown org as WARNING when not found via RHSM API
The InventoryController will now throw an ExternalServiceException
instead of a general ApiException. If a 400 is detected when listing
consumers from the RHSM API, it assumes that the org does not exist
and sets an appropriate error code.

Callers have been updated, where applicable, to log a warning instead
of an error (logging the entire exception).