Clean up error message that is masking the exception trace
Due to the use of multi-argument error logger it was treating the exception as
an argument to the formatter and not a throwable to be logged.
The MDC logger is already logging the org_id.
Also removed the call to get the account number which is no longer necessary as we
are doing all the calculations by org_id.