Throw proper exception when contracts are not found
Since the Contracts API retruns an empty list when a contract
is not found based on the search criteria, this should not
result in an ExternalServiceException. Instead we now throw
a ContractMissingException which is handled by the
BillableUsageController.

As part of this change, ContractController.getContractCoverage
now returns a Double instead of Optional, ensuring that if the
call succeeds, a value will be returned.