SWATCH-1257 Ingest contract dimensions from Partner API
Instead of ingesting the dimensions from UMB message we will use Partner API to populate dimensions into metrics. We will get those contracts that doesn't have an endDate since those will be active contracts. The old contracts with no endDates are tests so we can ignore it. Also if the dimensions are empty we will not do anything.