Add rollback for remittance history DB changes
A rollback changeset was created to revert the changes made in
/liquibase/202305231003-update-remittance-table-for-history.xml

If a rollback is required, a PR will need to be created that reverts
the remittance history logic and also enables the rollback changeset
in the changlog.

When the rollback completes, the remitted_value will be in billable
units (as if we were applying a billing factor of 1.0). Since the new
remittance logic would be reverted along with enabling the rollback,
the next time a tally summary is processed, the units will be converted
based on the configured billing factor and continue as usual.

Rollback Steps:
* Re-add removed version column (default 1)
* Re-add the billing_factor column (default 1.0)
* Drop the FK so that we do not get collisions when inserting summed
  records.
* Insert new records based on the sum of the remittance_pending_value
  grouping on PK values which will be in RAW format (billing factor: 1.0).
  The granularity will be set to 'MONTHLY' so that they can be
  distinguished. The remitted_pending_date will be the MAX date of the
  grouped remittance records.
* Delete all remittance records that are not 'MONTHLY' granularity (new)
* Revert the changed column names back to original values.
* Re-create the PK.
* Drop the granularity column.