Issue single DELETE query when purging Event records
By default, a derived delete query will return instances matching the where clause prior to deleting the entities one at a time.  This was too memory intensive.  Since we're not concerned with invoking any lifecycle callbacks, we can override the behavior with @Query to delete all matching records in bulk. 

https://docs.spring.io/spring-data/jpa/docs/current/reference/html/#jpa.modifying-queries.derived-delete