Make alteration in how HypervisorReportCategory is handled
Prior to the migration to the subscription_measurement table, a
hypervisor report category of "non-hypervisor" would return capacity
records that *had* physical sockets/cores and *did not have* hypervisor
sockets/cores.  This logic was easily implemented since every capacity
record had the combination of physical sockets, physical cores,
hypervisor sockets, and hypervisor cores.

In the subscription_measurement world, that tuple of attributes is now
spread out across four rows.  The logic to ensure that a grouping of
measurements for a subscription has physical metrics without having any
hypervisor metrics would be more complex and computationally expensive.
Based on discussions with the team lead, this patch changes the behavior
of the "non-hypervisor" report category to filter based on the presence
or absences of measurement records with a measurement type of
"physical".