Use consistent casing for metric ID names
The migration code was written to store the subscription_measurement
records with the metric ID as "CORES" or "SOCKETS".  However, in various
places, the code was using toString on the members of the MetricId enum.
This would be fine except the toString value is "Cores" or "Sockets" so
no measurements were matching!