Improve performance in SubscriptionTableController
The old code was fetching the offerings for the SKUs one at a time as we
moved through the subscriptions.  Instead of hitting the DB over and
over again for one offering at a time, this patch consolidates the
queries so that we fetch all the offerings at once.

See https://issues.redhat.com/browse/SWATCH-1343