Fix incorrect RH_ prefix from some internal roles
Description
===========

In some places we were granting `RH_INTERNAL` and in some places simply
`INTERNAL`. Because of this, it was not sufficient to auth via x509 cert
when using internal endpoints.

Testing
=======

Steps
-----

1. Deploy the service:

```shell
./gradlew :bootRun
```

2. Hit the opt-in endpoint w/ a header mimicing turnpike w/ x509 auth:

```shell
http PUT localhost:8000/api/rhsm-subscriptions/v1/internal/rpc/tally/opt-in \
  org_id==org123 \
  account_number==account123 \
  x-rh-identity:$(echo '{"identity":{"auth_type":"x509","type":"X509","x509":{"subject_dn":"CN=test.example.com"}}}' | base64 -w0) \
  Origin:console.redhat.com
```

With `main`, you'll get a 403, with this change, you'll get a success.