Truncate clock end dates to microseconds
Postgresql does timestamp comparison at microsecond precision,
and rounds up when a query is made using a timestamp with
microsecond precision. This can lead to unexpected query results.

The ApplicationClock was using LocalTime.MAX to force dates to
the max time for a day/month which results in microsecond precision,
which often gets passed as DB query parameters.

This change updates the clock to truncate the max time to microsecond
precision, preventing the rounding issue at the DB.

Before change: 2023-03-31T23:59:59.999999999Z
After change: 2023-03-31T23:59:59.999999000Z