Remove bidirectional relationship between Offering and Subscription
We never really needed Offering to maintain a list of Subscriptions.  I
added it because it made testing easier.  However, the cascade we had on
the Offering side was causing Hibernate to throw an
EntityExistsException in SubscriptionSyncController.syncSubscription.
Ultimately, Offering objects do not need to know about their
Subscriptions so this patch just removes that half of the relationship
and fixes the tests accordingly.