Make remittance updates immutable
To better handle remittance updates on re-tally, remittance records are
now immutable. It is less important that Swatch knows the details of what
snapshot caused remittance to occur, than it is to know that it has occured,
and when it happened.

This patch no longer updates existing remittance or tracks remittance
on a granular level. Whenever Swatch detects that it needs to remit usage,
a new remittance record is created and the remittance_pending_date is set
to the current timestamp (not the snapshot date). By doing so we only ever
bill for any additional usage that was detected during re-tally. Should a
re-tally result in less usage than what has already been remitted, usage
will not be billed for until it exceeds the already remitted amount.

Changes:
* Remove granularity column from table and field from BillableUsageRemittanceEntity
* Set remittance_pending_date to current timestamp instead of
  snapshot_date
* Only create a new remittance record if remitted_pending_value is > 0
* Continue to always send a BillableUsage message regardless of if the
  remitted value is 0.
* Remove granularity from remittance retention policy and instead use a
  single retention duration defaulting to 70 days (matching hourly
tally)
* Removed the remittance sync code since it was relying heavily on
  remittance granularity.
* Added some TEMPORARY code to the Tally API that will return
  equivellant results when contracts are in play.
  NOTE: This will not behave correctly when a re-tally happens. Tallies
        must occur hour by hour as they do naturally.