Remove erroneous trigger from subscription_capacity table
During the first phase of the migration to the subscription_measurements
table we had to migrate the has_unlimited_usage column from the
subscription_measurements table to the subscriptions table.  As part of
this migration, we added a trigger to the subscription_measurements
table so that subsequent changes to its has_unlimited_usage column would
be reflected in the subscription table.

We then added a foreign key relationship between the offering and
subscriptions table (SWATCH-1365).  Part of that effort moved the
has_unlimited_usage column from the subscriptions table to the offering
table.  In that card we dropped the has_unlimited_usage column off
subscription entirely since that column is also on the offering table.
However, I forgot to remove the triggers on subscription_capacity.
Those triggers were then left trying to insert data into a column that
no longer existed.

This patch deletes the incorrect trigger and procedure.