Invalidate the SKU existence cache properly when offerings are saved
During a subscription sync, the following events can occur:

* A check for a SKU is performed.  The SKU is not found
* A value of "false" for that SKU is written to the cache that stores
  the existence of SKUs
* We save an offering with that SKU
* Another subscription with the same SKU comes along and we make another
  call to OfferingRepository.existsById
* The cache returns false instead of true because we did not update the
  offering existence cache after saving the offering

This patch removes a SKU from the SKU existence cache after a save or
saveAndFlush operation.  For saveAll or saveAllAndFlush operations, the
entire cache is invalidated because there is no way to loop through the
iterable of saved entities to extract the SKU from each one.