Support importing the Contract's generated Capacity client
Because the capacity client is already being generated
for swatch-contracts and we are already importing the
contracts client, allow injecting the Capacity API as
well.

An HttpClient extension was added that will add the Red
Hat identity header to the request (if it exists and is
required) so that it is forwarded along to the Capacity
client's request.