Consider subscription as active when end_date is null
A subscription created via a contract may or may not have an end_date
set. When fetching capacity during a tally, subscriptions with a null
end date were being filtered out and not being included in the
capacity report and therefor not being applied to the tally's running
totals.

This patch updates made the the SubscriptionRepository to consider a
subscription as active when the end_date is null.