On running total tallies, report the past as has_data true
Right now on calls to /tally/products/{product_id} we are marking days
that don't have an associated measurement with has_data=false.  This
choice doesn't really make sense with a running total since even if we
don't have a measurement, we do still have data.  We should, however,
mark days in the future with has_data set to false.