Add order by clauses on SubscriptionRepository to avoid Hibernate issue
HHH-17040 is a Hibernate issue that can cause a single root entity to be
returned in duplicate depending on the row order of the result set.  We
can side-step the issue by adding an "order by" clause based on the
composite primary key we use for Subscription.  That "order by" will
ensure all rows of the same Subscription are grouped together.