Prevent TestConfiguration classes from polluting other tests
Within the Spring documentation on the TestConfiguration annotation is
the line "If you directly use @ComponentScan (that is, not through
@SpringBootApplication) you need to register the TypeExcludeFilter with
it. See the Javadoc for details."  We use ComponentScan extensively to
control how our configuration is imported.  Without the
TypeExclusionFilter, we were inadvertently loading TestConfiguration
annotated classes and those beans were being used across multiple tests.

I discovered this when I realized that
InventoryAccountUsageCollectorTallyTest was actually relying on
TestConfiguration declared in HourlySnapshotRollerTest.

This patch inserts the excludeFilters that are set on the
SpringBootApplication annotation so that our code will follow the
expected behavior of a Spring Boot project.