Replace tests needing prometheus service to integration tests
Instead of mocking the prometheus service in the following tests:
- PrometheusMeteringControllerTest
- PrometheusServiceTest

We can configure Wiremock to mock the Prometheus Query API. 
This way the tests will be more readable and also we can improve them to also mock large content of data. 

Note that we can only use this approach in tests that also use the SpringBootTest extension.