Perform running total rounding after all calculations instead of during
Suppose we have 4 snapshots all with values of 3.1.  Our old running
total calculation method would round each value up individually and then
total them for a running total series that looked like 4, 8, 12, 16.
Meanwhile, the monthly total calculation would add all the floating
points together and round up at the end for a value of 13 which does not
square with the final running total of 16.

This path performs running total calculations without rounding up and
defers the rounding until the data is ready to be transmitted to the
requester.  With this patch, the running total series will be 4, 7, 10,
13.