Change return to continue
In refactoring this code, a block moved from being a forEach closure to
being a for loop.  In a forEach closure, a return statement has the
effect of moving to the next item in the collection.  In a for loop, the
loop is short-circuited.  As a result, we were not doing all the tallies
we needed to do.