Prevent Resteasy from adding its BeanPostProcessor twice
Resteasy creates a BeanPostProcessor, org.jboss.resteasy.plugins.spring
.SpringBeanProcessor.ResteasyBeanPostProcessor that is responsible for
registering everything annotated with Provider.  The custom
ResteasyProviderRegistrar that we wrote to enable Resteasy in our tests,
was causing this postProcessor to be added to the list of postProcessors
twice.  The result was a warning in the start-up logs about not
registering providers twice.

This patch makes the bean in the ResteasyProviderRegistrar contingent on
the presence of a Mockito bean (which will only be there in tests).
It's a bit of a hack, but it was the best way I could figure out to
determine if we were in a test environment.