Remove the concept of retally during Hourly Tally
This patch completely removes the concept of an hourly retally
from the hourly tally process. When the hourly tally process runs,
Events are batch processed based on the Event.record_date,
instead of relying on a date range.

Events are fetched based on the last event that was tailled
and are processed in batches (configurable). A tally_state
table was added to track the last known Event.record_date
that was successfully processed, and is used to limit the
Event query when the tally operation occurs.

The API date range parameters have been removed. When the
hourly tally operation runs, it will process any events that
have not yet been processed.

As well, the overall hourly tally host and calculation process
has changed. Events with the same timestamp are now applied
accumulatively. For example, two events with the same
measurement for 10am will result in a 10am hourly snapshot
with a value equals to the sum of the measurements of both
events.

This makes it possible to perform ammendments on a Host's
monthly totals and affected account Snapshots directly,
by applying negative measurement values to a host.

In order to support retry-on-error via the RetryTemplate,
host updates via events needed to be idempotent. This was
accomplished by adding a lastAppledEventRecordDate to the
Host record. Since Events are processed in order by
record_date, this allows for determining if an Event had
already been processed successfully. Also, an
AccountUsageCalculationCache was created to track the
calculations that were computed as an Event batch is
processed. On retry, the event record date is again used
to determine if the Event was already applied to a
calculation.

NOTE:
* A default batch size of 16000 Events is used to keep memory
  usage at a reasonable level.
* When an account is reset, all TallyState records will be
  removed.