Do not send usage to AWS when outside configured window
The AWS Marketplace API will only accept usage in which
the timestamp must be on or after the START_OF_CURRENT_HOUR - 6h.

In order to avoid sending requests to the AWS API that we know will
fail, the AWS producer should check that the incoming usage is within
this window before sending the usage to AWS.

If the usage is outside of this window, the usage should not
be sent to AWS, a WARNING message should be logged, and an
'ignored' micrometer counter should be incremented.

An AWS_MARKETPLACE_USAGE_WINDOW configuration property was added to
allow us to adjust our pre-send window quickly if we ever need to.
The default value for this property is 6h.