Update promql templates to handle unexpected duplicate instance results
We expect a single cluster per metric.  In case of bad data/misconfiguration by the reporter, it's possible to have multiple values.  When this happens, prometheus throws an exception when trying to join.

In the event that the metric corresponds to multiple clusters/instances (identified by unique instance keys, generally _id), we'll take the max value of those to ensure one-to-many matches and avoid an attempt at a many-to-many match.

https://issues.redhat.com/browse/SWATCH-1771