Process Events in batches while streaming from DB
This patch streams EventRecords from the DB until a batch size
is reached, at which point the Events are passed along for
processing. This results in a small reduction in overall memory
usage, as well as a slight improvement in perfromance, while
processing a large number of Events at once.