Don't halt DB pod start-up on error
Any restart to the DB pod will result in an error since we try to create
the necessary user every time.  Rather than add the more complex logic
to detect whether the user exists or not, I'm just going to set the DB
start-up to keep going even if it encounters an error.