Use event_id as EventRecord PK
When spring-boot-dependancies was updated, one of the dep
changes was causing persistance exceptions because we were
attempting to update the PK in the @PrePersist method.

We no longer need to have the uniqueness constraint on the
events table since we support processing multiple Events for
the same timestamp. This allows us to use a UUID as the only
unique identifier instead of a now useless composite key.