Define 'rhelemeter' query template for els payg product definition
We're able to remove some of the complexity of the promql as defined by the default query template around the metadataMetric.

Instead of comparing the "product" label directly, we'll be using a regex in the promql.

For now, I'm opting to use the expression behind the recording rule in the "metric" field, due to some updates required to support recording rule evaluation times (as defined by SWATCH-2075)