Allow Spring Kafka configuration to operate using the system truststore
Under normal circumstances, if you wanted to use the system truststore,
you simply wouldn't set the kafka.ssl.truststore-certificates property
at all.  The KafkaProperties.Ssl instance would then have null for the
truststoreCertificates field and Kafka would initialize using the
default system truststore.

We, however, are using Clowder and we need to set the
truststore-certificates property if a certificate is provided to us.  We
use the ClowderJsonPathPropertySource class to do this and set the
property to something like `${clowder.kafka.brokers.cacert}`.
Unfortunately, if the ClowderJsonPathPropertySource can't resolve that
JSON path, the property ends up getting set to the literal string
"${clowder.kafka.brokers.cacert}" which of course is not a valid
certificate, nor does it indicate to Kafka to use the system truststore
which is what we would want to do if the Clowder JSON were silent about
the CA certificate to use.

We could use the value `${clowder.kafka.brokers.cacert:}` which would
resolve to the empty string since that is the default provided after the
colon.  However, the Kafka DefaultSslEngineFactory doesn't handle this
very well since getting an empty string is a pretty strange thing to get
for the value of a certificate.

This patch creates a bean post-processor that will change the truststore
properties back to null if any of them are set to the empty string and
consequently Kafka will use the system truststore.

See https://issues.redhat.com/browse/SWATCH-2088