Resolve incoming Event conflicts and amend differences
When processing incoming Events from either the topic or Admin API,
we now check if there are any events that are in conflict, and resolve
them when required.

A conflict occurs when there is an existing event that matches the
EventKey (org_id, event_type, event_source, instance_id, timestamp)
of the incoming event. In the case of a match, conflict resolution
must occur.

Example Conflict/Resolution scenarios include:

1. No Conflicts
 - Incoming event is valid and no resolution required.
 - An EventRecord is created from the incoming Event.
2. Conflicting event with equal measurements
 - Incoming Event is ignored.
 - Existing EventRecord remains as is with no updates.
3. Conflicting event with different measurements
 - EventRecord added with a negative accumulative measurement value
   (resetting the current the total for the timestamp to 0).
 - Incoming EventRecord measurements reflect the values at this
   timestamp.
  - e.g. An incoming event has a measurement of 10 cores. An existing EventRecord matching
         the tuple has a measurement of 20 cores. An EventRecord would be created with
         a cores value of -10 and another EventRecord created to apply the new value of 20
         cores.

NOTES:
1. Removed the CleanupEvent

We no longer need to send/process cleanup events since the EventConflictResolver will
address any existing events so that they are reflected when tallied.