SWATCH-2109 Redefine rhelemeter querytemplate PromQL
Added the breakdown explanation of the PromQL query to the swatch-metrics/README.md file.

We've opted to not use the recording rule for two reasons. First, it was only further complicating an already complicated promql query. Second, we get more accurate data by not using it. When we requested a recording rule to be made, we expected that its evaluation would happen the same frequency as the openshift telemeter instance. Since that's not going to be the case, using a recording rule doesn't benefit us much.

Because of the timing of metrics being reported to prometheus, and how the average value over the hour is calculated, it's possible that an instance has datapoints with different metadata labels that overlap. For this particular scenario described in the card, there was an overlap where there was instance metadata indicating Premium SLA and another datapoint for Standard SLA. The query that we landed on will take the first metadata label seen.

What this means functionally is if a metadata field is updated, it can take up to an hour for any change in support level or billing account for a system before it's reflected in prometheus. This can also result in us underbilling for the period of time between when the metadata field was changed and top of the next hour. To address underbilling, we'd have to get a recording rule to record max(system_cpu_logical_count) by (_id) and then have it run at least every 10m, and then replace system_cpu_logical_count in the numerator with it.