Set the usage date to now after remittance is created
Before this change, the remittance accumulation_period was getting
set to the current YYYY-MM, but should have been the month of the
incoming usage snapshot.

When performing a re-tally, existing remittance was being determined
based on the month of the snapshot, however, was getting recorded in
the DB as the current month. This would result in new remittance
getting created, and being billed for, every time a re-tally was run
that was missing remittance in a past month.