SWATCH-2162: Send metrics using backpressure mecanisms into Kafka
This solution has been agreed with Clement (the maintainer of the Quarkus messaging extensions). He suggested to use the emitter.hasRequests method to be sure the broker can handle more messages. The hasRequests works like a capacity barrier: when we send up to 1024 messages (1024 is the default value) without getting an ack from the broker, we wait until sending more messages).

Update: there seems to be two upper limits, (1) the one configured by the max-inflight-messages which is the one used by the hasRequests method and default value is 1024; and (2) the buffer limit which is used by the overflow strategy and default value is 128. Therefore, for this solution to work, we need to set the max-inflight-messages limit to the same as the buffer limit.

Relates: https://github.com/smallrye/smallrye-reactive-messaging/issues/2465