Use a separate client certificate for the UMB in the ephemerals
Since the ephemeral environments are short-lived, the UMB needs to
garbage collect the queues once the ephemeral environment goes inactive.
To do this collection, they need the ephemeral clients to have
"ephemeral" in the name.  We created a new service account,
nonprod-insightsrhsm-ephemeral, and issued a client certificate for it.

This commit makes the adjustments necessary to use that certificate
(before we were using a common client certificate across multiple
services).

It also moves the spring resource prefix from template var to
environment variable.  This keystore path is shared between Spring and
Quarkus applications.  If overridden (as it is in the ephemeral
environments), it's defined as a path without the Spring specific
resource prefix "file:".  So we need to remove that prefix from the
template value which will be override and down into the environment
variable value.

See SWATCH-1044