Add retries/debug logs to Partner Gateway processing
Jira issue: [SWATCH-2206](https://issues.redhat.com/browse/SWATCH-2206)

Description
===========

In order to handle transient issues with the data coming back from the partner gateway API, as well
as provide debug logging to assist the partner gateway engineers in debugging issues.

Note that I added a custom debug logger `DebugClientLogger` that can be configured to log the
request and response bodies. See comments in `application.properties` for details. (It's configured
for partnerEntitlement requests for `dev`, `test`, and `stage` profiles only).

I changed the contract entity validation to use hibernate validator, to get better messages about
which fields are missing. (Each failed field is now logged at a `WARN` level).

I tweaked the API enrichment so that if the contract fails validation afterward, the enrichment
will be retried up to 10 times, with a 500ms delay between each retry (adding ~5 seconds total for
this unexpected case). Given we expect the initial message (not having an orgId for the contract)
to fail until the contract is linked, this type of failure is _not_ retried.

Testing
=======

Run the AzureContractLifecycleIntegrationTest to see changes in logging and retries:

```shell
./gradlew :swatch-contracts:test -i --tests=AzureContractLifecycleIntegrationTest
```

Observe:
* Retries are verified by wiremock `verify` calls.
* Request and response bodies are logged (look for `DebugClientLogger` logs in the output).