Include metrics without "support" labels when querying rhelemeter
Prometheus treats an unset label the same as an empty string. Adding a | at the end of the support regex will make the query return results when there isn't a support label set. When the support label is unset, it should that usage should default to "Premium", as defined in the rhel-for-x86-els-payg product config yaml.