Change Splunk certificate verification in dev to false
When using Splunk in the dev profile, developers will be sending their
events to a Splunk docker container that will not have a server
certificate signed by a recognized certificate authority.  Instead the
certificate will be self-signed.  Configuring the applications to
disable certificate validation in the dev (and only in the dev) profile
will be make life easier for developers while still maintaining security
in non-development environments.