Fix 404 issue when POSTing to export service API
There are two UUIDs that export service cares about when POSTing to their /upload and /error endpoints.  An export id and an export resource id.

We had the two parameters swapped when calling downloadExportError and downloadExportUpload.  Updated that and then renamed some of our local variables to hopefully prevent further confusion.