Enable Liquibase in init containers
The init containers are meant to be running Liquibase and the main
containers are not.  Consequently, the main containers disable Liquibase
via an environment variable.  The problem is that the init containers
inherit their environment from the main container and disable Liquibase
as well.  The init container, therefore, does nothing at all!

This patch enables Liquibase explicitly in the init containers.