Update unit test to use persisted object as base of comparison
Despite appearances to the contrary, SimpleJpaRepository.save() actually
ends up calling a `merge` rather than a `persist` in this unit test.
Hibernate HHH-17634 changed Hibernate's behavior (to match the JPA
spec) such that after a merge a deep copy of an object is copied into
the managed entity and the original object is left unmodified.

In this test, that meant that the HostBucketKey on the original Host was
previously being updated with the generated hostId but after the update
to Hibernate 6.4.4 in Spring 3.2.3, the HostBucketKey was no longer
being updated since a deep copy of the entity graph is made instead.