SWATCH-2195: Do not wait for hourly tally operation to finish when synchronous_request is true
Currently when we are trying to run hourly tally operation on large orgs with synchronous_request: true, we are seeing 504 Server Error: Gateway Time-out for url. We just need a way to skip queues and not to wait for response till operation finish to avoid 504 gateway timeouts.

The idea is to reuse the executor task queue that was in place along with the kafka task queue. 
Before these changes, only one task queue was started (when using the kafka-queue profile, it was reading tasks from the topic). After these changes, the in memory task queue will always be started and will be used when running the hourly tally operation when the header "x-rh-swatch-synchronous-request:true" is provided.