downgrade postgresql instance to 12
I had to downgrade the version we're using of postgresql to use the same version than in the stage,prod and ephemeral environments. Before, we were using Postgresql 13 which had some functions to generate UUID that are not by default in Postgresql 12, so I could not spot this when building locally. Using the Postgresql 12 always is safer.