Enable --select-1 on pod and project selection
A fix in iterfzf means we can use the --select-1 fix which allows us to
bypass an interactive fzf flow if the initial query only matches one
item in the list.  Now developers can just write a specific enough
invocation using the --pod-prefix and --project options and they won't
have to endure an interactive process.  The interactive process is still
nice for when one is first starting, but at a certain point during
development, the project and pod to deploy to are known quantities and
constantly having to select them is tedious.