Separate mutable-jar aspect from dev profile
Rather than relying on the dev profile for building a mutable-jar, I
want to enable this property explicitly and separately.  I can foresee a
circumstance in which we have dev profile build options that are
desirable for local deployment but that we don't want for ephemeral
deployments.  Coupling the hot reload aspect of the code with the dev
profile seems like an unnecessary encumbrance.