Ensure consistent artifact version names for hot deployment
The default version used for built artifacts has the git hash in it.  If
you do a git commit during the course of development, the file name of
the artifact changes.  And if the file name changes, instead of
replacing your old artifact, rsync just copies over the new file and a
hot reload is never triggered since the old artifact doesn't change.