SWATCH-2387: Update tag on postgresql-12-centos7 database (#3198)
<!-- Replace XXXX with the issue number. Issue will be auto-linked -->
Jira issue: SWATCH-2387

## Description
<!-- Provide a description of this PR. Try to provide answers to "what",
"how",
and "why" -->
The `latest` tag was removed on this container as the container image is
no longer maintained. The patch adds the `centos7` tag as a stop-gap
until we can upgrade.

This issue is what's causing the floorist query validation task to fail.

## Testing
<!--
When possible, please use commands a developer can directly paste or
modify
-->

### Setup
1. **Be sure you are willing to destroy your database**. Do a `pg_dump`
if you are not so you can import later.
1. `podman-compose down`

### Steps
<!-- Enter each step of the test below -->
1. `podman-compose up -d`

### Verification
<!-- Enter the steps needed to verify the test passed -->
1. You'll see the `quay.io/centos7/postgresql-12-centos7:centos7` image
get pulled.